refactor(modules): extract shared-kernel bounded context
First step of v4 reorganization into src/modules/<bc>/ — moves the DDD
Shared Kernel types depended on by multiple bounded contexts.

Moves (with history preserved via git mv):
- src/entities/shared/         → src/modules/shared-kernel/entities/shared/
- src/entities/diffStats/      → src/modules/shared-kernel/entities/diffStats/
- src/entities/language/       → src/modules/shared-kernel/entities/language/

26 importers updated to the new @/modules/shared-kernel/... paths.
Also normalized 2 relative imports to the @/ alias (project rule:
no relative paths).

yarn typecheck: zero new errors (only the preexisting 'join' unused
warning in dashboardModulesCoverage.acceptance.test.ts).
