+++ src/entities/tracking/reviewEvent.ts
@@ -1,4 +1,4 @@
-import type { DiffStats } from '@/entities/diffStats/diffStats.js';
+import type { DiffStats } from '@/modules/shared-kernel/entities/diffStats/diffStats.js';
 
 export interface ReviewEvent {
   type: 'review' | 'followup';
