+++ src/frameworks/queue/pQueueAdapter.ts
@@ -2,7 +2,7 @@ import PQueue from 'p-queue';
 import type { Logger } from 'pino';
 import { loadConfig } from '@/frameworks/config/configLoader.js';
 import type { ReviewProgress, ProgressEvent } from '@/entities/progress/progress.type.js';
-import type { Language } from '@/entities/language/language.schema.js';
+import type { Language } from '@/modules/shared-kernel/entities/language/language.schema.js';
 import type { ClaudeModelName } from '@/entities/modelRouting/modelRouting.schema.js';
 
 export interface ReviewJob {
