+++ src/interface-adapters/controllers/webhook/gitlab.controller.ts
@@ -29,7 +29,7 @@ import { startWatchingReviewContext, stopWatchingReviewContext } from '@/main/we
 import type { ReviewContextGateway } from '@/entities/reviewContext/reviewContext.gateway.js';
 import type { ThreadFetchGateway } from '@/entities/threadFetch/threadFetch.gateway.js';
 import type { DiffMetadataFetchGateway } from '@/entities/diffMetadata/diffMetadata.gateway.js';
-import type { DiffStatsFetchGateway } from '@/entities/diffStats/diffStatsFetch.gateway.js';
+import type { DiffStatsFetchGateway } from '@/modules/shared-kernel/entities/diffStats/diffStatsFetch.gateway.js';
 
 export function extractBaseUrl(remoteUrl: string): string | null {
   try {
