+++ src/modules/shared-kernel/entities/diffStats/diffStatsFetch.gateway.ts
@@ -1,4 +1,4 @@
-import type { DiffStats } from '@/entities/diffStats/diffStats.js';
+import type { DiffStats } from '@/modules/shared-kernel/entities/diffStats/diffStats.js';
 
 export interface DiffStatsFetchGateway {
   fetchDiffStats(projectPath: string, mergeRequestNumber: number): DiffStats | null;
