+++ src/modules/shared-kernel/entities/shared/duration.valueObject.ts
