+++ src/tests/factories/diffStats.factory.ts
@@ -1,4 +1,4 @@
-import type { DiffStats } from '@/entities/diffStats/diffStats.js';
+import type { DiffStats } from '@/modules/shared-kernel/entities/diffStats/diffStats.js';
 
 export class DiffStatsFactory {
   static create(overrides: Partial<DiffStats> = {}): DiffStats {
