+++ src/tests/factories/projectStats.factory.ts
@@ -1,5 +1,5 @@
 import type { ProjectStats, ReviewStats } from '@/entities/stats/projectStats.js';
-import type { DiffStats } from '@/entities/diffStats/diffStats.js';
+import type { DiffStats } from '@/modules/shared-kernel/entities/diffStats/diffStats.js';
 
 export class ReviewStatsFactory {
   static create(overrides: Partial<ReviewStats> = {}): ReviewStats {
