+++ src/tests/units/entities/duration.valueObject.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { Duration } from '../../../entities/shared/duration.valueObject.js';
+import { Duration } from '@/modules/shared-kernel/entities/shared/duration.valueObject.js';
 
 describe('Duration', () => {
   describe('factory methods', () => {
