+++ src/usecases/stats/backfillDiffStats.usecase.ts
@@ -1,5 +1,5 @@
 import type { StatsGateway } from '@/entities/stats/stats.gateway.js';
-import type { DiffStatsFetchGateway } from '@/entities/diffStats/diffStatsFetch.gateway.js';
+import type { DiffStatsFetchGateway } from '@/modules/shared-kernel/entities/diffStats/diffStatsFetch.gateway.js';
 import type { BackfillProgress } from '@/entities/backfill/backfillProgress.js';
 
 export interface BackfillDiffStatsDependencies {
